pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-manager"
version = "0.1.0"
description = "Multi-Agent Coordination/Planning (MCP) Server using FastAPI and SQLAlchemy ORM"
readme = "README.md"
authors = [
    {name = "Agent Manager Team"},
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "pydantic>=2.4.0",
    "openai>=1.3.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "asyncpg>=0.28.0",
    "aiosqlite>=0.19.0",
    "httpx>=0.25.0",
    "python-multipart>=0.0.6",
    "redis>=5.0.0",
    "hiredis>=2.2.0",
    "prometheus_client>=0.19.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "flake8>=6.0.0",
    "pre-commit>=3.4.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]